Change loop count
The test is designed to expire once it reaches the end of a file, therefore the number of threads and loops only need to exceed any likely scenario. For now it's set to 20 threads and 1000 loops, giving an import file of 20000 records. This is impossible at the moment due to an app limit of <15000.