Replace `--browser_or_device` option
This replaces the command line option `--browser_or_device` with the
following three options:
- `--browser`
- `--browser-channel`
- `--device`

The reason for this is to align this project with the default command
line options used by `pytest-playwright` to bring us closer to being
able to use this plugin in the future (and simplify most of our
Playwright configuration code).

https://playwright.dev/python/docs/test-runners#cli-arguments

It also means we don't need to maintain our own list of browsers and
devices, which gives us more flexibility allowing the tests to be run
in various different combinations easily without needing to make a
change to the code.