Add `--slowmo` option
This adds a new `--slowmo` command line option when running the tests to
make it easy to re-run the tests with a slow motion parameter without
needing to set a new environment variable.

This also brings us more in alignment with the command line variables
for `pytest-playwright` which will make it easier for us to migrate to
that in the future.

https://playwright.dev/python/docs/test-runners#cli-arguments