Add `--headed` option
This adds a new `--headed` command line option when running the tests
to make it easy to run the tests in a headed mode when request.

This also brings us more in alignment with the command line variables
for `pytest-playwright` which will make it easier for us to migrate to
that in the future.

https://playwright.dev/python/docs/test-runners#cli-arguments