Make `reset_environment` a fixture
This allows us to decouple the method from the singleton class
`CurrentExecution` which makes it more flexible (it can itself depend on
other fixtures, and by avoid global state we can work towards making
these tests parallelizable).

This is also a requirement to switching to pytest-playwright, which
requires us to avoid a global state object (`CurrentExecution`) and
replace it with the use of fixtures.