Add `--skip-reset` option
This replaces the `RESET_ENV_BEFORE_EXECUTION` environment variable to
make it easier to run quickly by having it available as a command line
argument.

By default, the reset endpoint will be called as part of the tests, but
when running with `--skip-reset` it avoids the reset endpoint being
called.