Add `base_url` fixture
This moves the logic related to the base URL in to a fixture, allowing
it to be used in tests directly without needing to go via the
`CurrentExecution` global singleton object. This will allow us to
migrate to using the `pytest-playwright` plugin in the future avoiding
the need for us to manage Playwright itself in our code.

I've also set up the contexts with a `base_url` allowing us to call
`page.goto("/")` without needing the full URL, which also allows us to
remove the `go_to_url` method which was just calling `page.goto`.