Add fixtures for pages
This adds a fixture for each page model object allowing it to be
included in each test that uses a particular page. This is necessary to
support removing the global `CurrentExecution` class as eventually these
fixtures will depend on the `page` fixture from pytest-playwright and
pass that in to the classes directly.