Refactor vaccines in to an enum
This refactors how vaccines are modelled to make it an enum object
that can be converted to a string so it can be used directly in the
tests, avoiding the need to check the vaccine value and allowing us to
add new vaccines easily in the future. I've also removed the coupling
between the name of the vaccine and the order in the page allowing us to
change the ordering without breaking these tests.