Refactor start page
This adds a new page model object, `StartPage`, which represents the
various start pages of the service (the main one for nurses and the one
for parents). By having this class we can refactor how many of the tests
are set up to improve readability, and also set up automatic screenshot
creation on each step.