Consent journey workflow
Additional workflow to test consent in isolation. As consent is now dependent on cohort data, the create cohort and upload cohort steps are included in this workflow. This differs from the other workflows in that it's designed to test consent only, so the input values are related to the consent journey rather than nurse journey.