Add `organisation` fixture
This adds a fixture which captures the name and ODS code of the
organisation, replacing the hard coded `test_data_values.ORG_CODE`. The
idea behind having this as a fixture means it can change dynamically,
and we can change the way the tests run to create a fresh organisation
each time.