Create organisation for each test run
This updates how the organisations are set up to ensure that a new
organisation is onboarded for each test run, and reset between each
module. This should allow test runs to execute in parallel, and avoid
issues where test date is shared between runs.

I've also had to refactor a few of the tests to ensure they can be run
in different organisations without breaking.