Remove waiting
This removes artifical waiting and the slow motion parameter to speed up
the test run. Locally I'm able to get the tests to run in about 10
minutes compared to 35 minutes previously.

To make this change, a number of refactors have been made to use
`wait_for` and `expect` which automatically handles waiting for the
elements to be visible without us needing to.