Reduce number of parallel test jobs
With the tests themselves running in parallel, running the devices in
parallal can lead to high numbers of failures due to timeouts or
incorrect database state. I don't think this job needs to run as fast as
locally so we can afford to slow down the number of parallel jobs.