Refactor vaccine pages
This refactors the vaccine page object in to separate page objects for
each distinct page in the service to ensure the classes and smaller and
don't leak elements from other pages.

I've also added a test to ensure the batch names are at least 2
characters long.

This change depends on a few changes to the service being deployed
first, specifically the improved accessibility labels:
https://github.com/nhsuk/manage-vaccinations-in-schools/pull/3717