Tidy up for consent questions
Different vaccinations have different questions associated with them, this was all hard coded in the original script. This makes it very flaky if a different combination of questions is presented. This simplification works with any number of consent questions. The next step will be to make it consistent across all vaccinations (and by extension make it ready for flu).