New jmeter script for large orgs
This script uses a predefined large org (>400 schools, >160K patients) to replicate the current largest org. It's designed to split the workload across schools, so performance issues seen with large number of patients in one school aren't seen. The difficulty is in the data source, so this test also employs a jmeter simple table server.

The corresponding GHA yaml is to be submitted separately so this can stay in branch until it's fully tested.