Support locations with a site
In https://github.com/nhsuk/manage-vaccinations-in-schools/pull/4297 we
added support for multiple distinct locations to exist at the same URN
by introducing an optional site code that is appended to the end of the
URN when identifying the location. This updates the tests to support the
possibility to working with a school that has a site code.